% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variables.R
\name{Apply.RV}
\alias{Apply.RV}
\title{Transform a random variable by a function.}
\usage{
\method{Apply}{RV}(self, func)
}
\arguments{
\item{func:}{function to Apply to the random variable}
}
\description{
Transform a random variable by a function.
}
\examples{
X = RV(Exponential(1))
Y = X.Apply(log)

Note: For most standard functions, you can Apply the function to
the random variable directly. For example, in the example above,
Y = log(X) would have been equivalent and more readable.

User defined functions can also be applied.

g <- function(x)
  return log(x ** 2)
Y = X \%>\% Apply(g)
}
