% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Defines a probability space for a Bernoulli
distribution.}
\usage{
Bernoulli(p)
}
\arguments{
\item{p}{(double) probability (number between 0 and 1)
of a "success" (i.e., 1)}
}
\description{
Defines a probability space for a Bernoulli
distribution.
}
