% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Beta}
\alias{Beta}
\title{Defines a probability space for a beta distribution.}
\usage{
Beta(a = 1, b = 1, xmin = 0, xmax = 1, scale = NULL)
}
\arguments{
\item{a}{(double): alpha parameter for beta distribution}

\item{b}{(double): beta parameter for beta distribution}
}
\description{
Defines a probability space for a beta distribution.
}
