% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Binomial}
\alias{Binomial}
\title{Defines a probability space for a binomial distribution.}
\usage{
Binomial(n, p)
}
\arguments{
\item{n}{(int): number of trials}

\item{p}{(float): probability (number between 0 and 1)
that each trial results in a "success" (i.e., 1)}
}
\description{
Defines a probability space for a binomial distribution.
}
