% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{BoxModel}
\alias{BoxModel}
\title{Defines a probability space from a box model.}
\usage{
BoxModel(box, size = 1, replace = TRUE, probs = NULL,
  order_matters = TRUE)
}
\arguments{
\item{box, }{size, replace, probs, order_matters}
}
\value{
A BoxModel
}
\description{
Defines a probability space from a box model.
}
