% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ChiSquare}
\alias{ChiSquare}
\title{Defines a probability space for chi-square distribution.}
\usage{
ChiSquare(df)
}
\arguments{
\item{df}{(int): degrees of freedom}
}
\description{
Defines a probability space for chi-square distribution.
}
