% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{count.Results}
\alias{count.Results}
\title{counts the number of outcomes that satisfied a given criterion.}
\usage{
\method{count}{Results}(self, fun = function(x) TRUE)
}
\arguments{
\item{fun}{(outcome -> bool): A function that
takes in an outcome and returns a
True / False. Only the outcomes that
return True will be counted.}
}
\value{
int: The number of outcomes for which
the function returned True.
}
\description{
counts the number of outcomes that satisfied a given criterion.
}
