% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{DeckOfCards}
\alias{DeckOfCards}
\title{Defines the probability space for drawing from a deck of cards.}
\usage{
DeckOfCards(size = 1, replace = FALSE, order_matters = TRUE)
}
\arguments{
\item{size}{(int): How many draws to make.}

\item{replace}{(bool): Sample with replacement or without?}
}
\description{
Defines the probability space for drawing from a deck of cards.
}
