% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DiscreteUniform}
\alias{DiscreteUniform}
\title{Defines a probability space for a discrete uniform distribution.}
\usage{
DiscreteUniform(a = 0, b = 1)
}
\arguments{
\item{a}{(int): lower bound for possible values}

\item{b}{(int): upper bound for possible values}
}
\description{
Defines a probability space for a discrete uniform distribution.
}
