% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Bernoulli}
\alias{draw.Bernoulli}
\title{A function that takes no arguments and
returns a single draw from the Bernoulli distribution.}
\usage{
\method{draw}{Bernoulli}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Bernoulli distribution.
}
