% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{Draw.BoxModel}
\alias{Draw.BoxModel}
\title{A function that takes no arguments and returns a value(s) from the 
"box" argument of the BoxModel.}
\usage{
\method{Draw}{BoxModel}(self)
}
\description{
Based on BoxModel inputs: 
Number of values returned depends on the input of the "size" 
argument. 
Whether or not a value in the box can appear multiple times
depends on the "replace" argument.
If a list of probabilities is specified, values drawn will be drawn
with the specified probabilities.
}
