% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Cauchy}
\alias{draw.Cauchy}
\title{A function that takes no arguments and
returns a single draw from the Cauchy distribution.}
\usage{
\method{draw}{Cauchy}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Cauchy distribution.
}
