% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.ChiSquare}
\alias{draw.ChiSquare}
\title{A function that takes no arguments and
returns a single draw from the ChiSquare distribution.}
\usage{
\method{draw}{ChiSquare}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the ChiSquare distribution.
}
