% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Exponential}
\alias{draw.Exponential}
\title{A function that takes no arguments and
returns a single draw from the Exponential distribution.}
\usage{
\method{draw}{Exponential}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Exponential distribution.
}
