% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.F}
\alias{draw.F}
\title{A function that takes no arguments and
returns a single draw from the F distribution.}
\usage{
\method{draw}{F}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the F distribution.
}
