% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Gamma}
\alias{draw.Gamma}
\title{A function that takes no arguments and
returns a single draw from the Gamma distribution.}
\usage{
\method{draw}{Gamma}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Gamma distribution.
}
