% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Geometric}
\alias{draw.Geometric}
\title{A function that takes no arguments and
returns a single draw from the Geometric distribution.}
\usage{
\method{draw}{Geometric}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Geometric distribution.
}
