% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Hypergeometric}
\alias{draw.Hypergeometric}
\title{A function that takes no arguments and
returns a single draw from the Hypergeometric distribution.}
\usage{
\method{draw}{Hypergeometric}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Hypergeometric distribution.
}
