% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.NegativeBinomial}
\alias{draw.NegativeBinomial}
\title{A function that takes no arguments and
returns a single draw from the Negative Binomial distribution.}
\usage{
\method{draw}{NegativeBinomial}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Negative Binomial distribution.
}
