% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Pascal}
\alias{draw.Pascal}
\title{A function that takes no arguments and
returns a single draw from the Negative Binomial distribution.}
\usage{
\method{draw}{Pascal}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Negative Binomial distribution.
}
