% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Poisson}
\alias{draw.Poisson}
\title{A function that takes no arguments and
returns a single draw from the Poisson distribution.}
\usage{
\method{draw}{Poisson}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Poisson distribution.
}
