% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variables.R
\name{draw.RVConditional}
\alias{draw.RVConditional}
\title{A function that takes no arguments and returns a value from
the conditional distribution of the random variable.}
\usage{
\method{draw}{RVConditional}(self)
}
\description{
A function that takes no arguments and returns a value from
the conditional distribution of the random variable.
}
\examples{
X, Y = RV(Binomial(10, 0.4) ** 2)
(X | (X + Y == 5)).draw() might return a value of 4, for example.
}
