% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.StudentT}
\alias{draw.StudentT}
\title{A function that takes no arguments and
returns a single draw from the T distribution.}
\usage{
\method{draw}{StudentT}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the T distribution.
}
