% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Uniform}
\alias{draw.Uniform}
\title{A function that takes no arguments and
returns a single draw from the Uniform distribution.}
\usage{
\method{draw}{Uniform}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Uniform distribution.
}
