% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Exponential}
\alias{Exponential}
\title{Defines a probability space for an exponential distribution.
Only one of scale or rate should be set. (The scale is the
inverse of the rate.)}
\usage{
Exponential(rate = 1, scale = NULL)
}
\arguments{
\item{rate}{double; rate parameter for gamma distribution
(often symbolized lambda)}

\item{scale}{double; scale parameter for gamma distribution
(often symbolized beta = 1 / lambda)}
}
\description{
Defines a probability space for an exponential distribution.
Only one of scale or rate should be set. (The scale is the
inverse of the rate.)
}
