% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Fdist}
\alias{Fdist}
\title{Defines a probability space for F distribution.}
\usage{
Fdist(dfN, dfD)
}
\arguments{
\item{dfN}{(int): degrees of freedom in the numerator}

\item{dfD}{(int): degrees of freedom in the denominator}
}
\description{
Defines a probability space for F distribution.
}
