% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{Filter.Results}
\alias{Filter.Results}
\title{Filters the results of a simulation and
returns only those outcomes that satisfy
a given criterion.}
\usage{
\method{Filter}{Results}(self, fun)
}
\arguments{
\item{fun}{(outcome -> bool): A function that
takes in an outcome and returns a
True / False. Only the outcomes that
return True will be kept; the others
will be filtered out.}
}
\value{
Results: Another Results object containing
only those outcomes for which the function
returned True.
}
\description{
Filters the results of a simulation and
returns only those outcomes that satisfy
a given criterion.
}
