% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Gamma}
\alias{Gamma}
\title{Defines a probability space for a gamma distribution.
Only one of scale or rate should be set. (The scale is the
                                          inverse of the rate.)}
\usage{
Gamma(shape, rate = 1, scale = NULL)
}
\arguments{
\item{shape}{(double): shape parameter for gamma distribution
(often symbolized alpha)}

\item{rate}{(double): rate parameter for gamma distribution
(often symbolized lambda)}

\item{scale}{(double): scale parameter for gamma distribution
(often symbolized beta = 1 / lambda)}
}
\description{
Defines a probability space for a gamma distribution.
Only one of scale or rate should be set. (The scale is the
                                          inverse of the rate.)
}
