% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Geometric}
\alias{Geometric}
\title{Defines a probability space for a geometric
distribution (which represents the number
of trials until the first success), including
the success.}
\usage{
Geometric(p)
}
\arguments{
\item{p}{(double): probability (number between 0 and 1)
that each trial results in a "success" (i.e., 1)}
}
\description{
Defines a probability space for a geometric
distribution (which represents the number
of trials until the first success), including
the success.
}
