% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Hypergeometric}
\alias{Hypergeometric}
\title{Defines a probability space for a hypergeometric
distribution (which represents the number of
ones in n draws without replacement from a box
containing zeros and ones)}
\usage{
Hypergeometric(n, N0, N1)
}
\arguments{
\item{n}{(int): number of draws (without replacement)
from the box}

\item{N0}{(int): number of 0s in the box}

\item{N1}{(int): number of 1s in the box}
}
\description{
Defines a probability space for a hypergeometric
distribution (which represents the number of
ones in n draws without replacement from a box
containing zeros and ones)
}
