% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{LogNormal}
\alias{LogNormal}
\title{Defines a probability space for a Log-Normal distribution
If Y has a LogNormal distribution with parameters mu and sigma, then
log(Y) has a normal distribution with mean mu and sd sigma.}
\usage{
LogNormal(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{(double): mean of the underlying normal distribution}

\item{sigma}{(double): standard deviation of the underlying normal distribution}
}
\description{
Defines a probability space for a Log-Normal distribution
If Y has a LogNormal distribution with parameters mu and sigma, then
log(Y) has a normal distribution with mean mu and sd sigma.
}
