% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{NegativeBinomial}
\alias{NegativeBinomial}
\title{Defines a probability space for a negative
binomial distribution (which represents the
number of trials until r successes), including
the r successes.}
\usage{
NegativeBinomial(r, p)
}
\arguments{
\item{r}{(int): desired number of successes}

\item{p}{(double): probability (number between 0 and 1)
that each trial results in a "success" (i.e., 1)}
}
\description{
Defines a probability space for a negative
binomial distribution (which represents the
number of trials until r successes), including
the r successes.
}
