% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Normal}
\alias{Normal}
\title{Defines a probability space for a normal distribution.}
\usage{
Normal(mean = 0, sd = 1, var = NULL)
}
\arguments{
\item{mean}{(double): mean parameter of the normal distribution}

\item{sd}{(double): standard deviation parameter of the normal
distribution (if specified, var parameter will be ignored)}

\item{var}{(double): variance parameter of the normal distribution}
}
\description{
Defines a probability space for a normal distribution.
}
