% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Pascal}
\alias{Pascal}
\title{Defines a probability space for a Pascal
distribution (which represents the number
of trials until r successes), not including
the r successes.}
\usage{
Pascal(r, p)
}
\arguments{
\item{r}{(int): desired number of successes}

\item{p}{(double): probability (number between 0 and 1)
that each trial results in a "success" (i.e., 1)}
}
\description{
Defines a probability space for a Pascal
distribution (which represents the number
of trials until r successes), not including
the r successes.
}
