% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Poisson}
\alias{Poisson}
\title{Defines a probability space for a Poisson distribution.}
\usage{
Poisson(lam)
}
\arguments{
\item{lam}{(float): rate parameter for the Poisson distribution}
}
\description{
Defines a probability space for a Poisson distribution.
}
