% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{ProbabilitySpace}
\alias{ProbabilitySpace}
\title{Defines a probability space.}
\usage{
ProbabilitySpace(drawFunc)
}
\arguments{
\item{draw}{function}
}
\value{
ProbabilitySpace
}
\description{
Defines a probability space.
}
