% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variables.R
\name{RVConditional}
\alias{RVConditional}
\title{Defines a random variable conditional on an event.}
\usage{
RVConditional(RV)
}
\arguments{
\item{random_variable}{(RV): the random variable whose conditional
distribution is desired}

\item{condition_event}{(Event): the event to condition on}
}
\description{
RVConditionals are typically produced when you condition a
RV on an Event object.
}
\examples{
X, Y = RV(Binomial(10, 0.4) ** 2)
(X | (X + Y == 5)) \%>\% draw() # returns a value between 0 and 5.
}
