% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{Sim.ProbabilitySpace}
\alias{Sim.ProbabilitySpace}
\title{Simulate n Draws from probability space.}
\usage{
\method{Sim}{ProbabilitySpace}(self, n)
}
\arguments{
\item{n}{(int): How many Draws to make.}
}
\value{
A vector containing the Simulation results.
}
\description{
Simulate n Draws from probability space.
}
