% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{StudentT}
\alias{StudentT}
\title{Defines a probability space for Student's t distribution.}
\usage{
StudentT(df)
}
\arguments{
\item{df}{(int): degrees of freedom}
}
\description{
Defines a probability space for Student's t distribution.
}
