% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Uniform}
\alias{Uniform}
\title{Defines a probability space for a uniform distribution.}
\usage{
Uniform(a = 0, b = 1)
}
\arguments{
\item{a}{(double): lower bound for possible values}

\item{b}{(double): upper bound for possible values}
}
\description{
Defines a probability space for a uniform distribution.
}
