% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{Apply.Results}
\alias{Apply.Results}
\title{Apply a function to each outcome of a simulation.}
\usage{
\method{Apply}{Results}(self, fun, ...)
}
\arguments{
\item{fun:}{A function to apply to each outcome.}
}
\value{
Results: A Results object of the same length,
where each outcome is the result of applying
the function to each outcome from the original
Results object.
}
\description{
Apply a function to each outcome of a simulation.
}
