% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Binomial}
\alias{draw.Binomial}
\title{A function that takes no arguments and
returns a single draw from the Binomial distribution.}
\usage{
\method{draw}{Binomial}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Binomial distribution.
}
