% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.DiscreteUniform}
\alias{draw.DiscreteUniform}
\title{A function that takes no arguments and
returns a single draw from the Discrete Uniform distribution.}
\usage{
\method{draw}{DiscreteUniform}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Discrete Uniform distribution.
}
