% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw.Normal}
\alias{draw.Normal}
\title{A function that takes no arguments and
returns a single draw from the Normal distribution.}
\usage{
\method{draw}{Normal}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Normal distribution.
}
