% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Draw.Uniform}
\alias{Draw.Uniform}
\title{A function that takes no arguments and
returns a single draw from the Uniform distribution.}
\usage{
\method{Draw}{Uniform}(self)
}
\description{
A function that takes no arguments and
returns a single draw from the Uniform distribution.
}
