% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{ProbabilitySpace}
\alias{ProbabilitySpace}
\title{Defines a probability space.}
\usage{
ProbabilitySpace(DrawFunc, self = NULL, other = NULL)
}
\arguments{
\item{draw}{function: A function explaining how to draw one
outcome from the probability space.}
}
\value{
ProbabilitySpace
}
\description{
Defines a probability space.
}
