% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variables.R
\name{RV}
\alias{RV}
\title{Defines a random variable.
 A random variable is a function which maps an outcome of
 a probability space to a number.  Simulating a random
 variable is a two-step process: first, a Draw is taken
 from the underlying probability space; then, the function
 is applied to that Draw to obtain the realized value of
 the random variable.}
\usage{
RV(probSpace, fun = function(x) x)
}
\arguments{
\item{probSpace}{(ProbabilitySpace): the underlying
probability space of the random variable.}

\item{fun}{(function, optional)}
}
\value{
A RV
}
\description{
Defines a random variable.
 A random variable is a function which maps an outcome of
 a probability space to a number.  Simulating a random
 variable is a two-step process: first, a Draw is taken
 from the underlying probability space; then, the function
 is applied to that Draw to obtain the realized value of
 the random variable.
}
\examples{
A single Draw is a sequence of 0s and 1s
, e.g., (0, 0, 1, 0, 1). P = BoxModel([0, 1], size=5)
}
