% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_space.R
\name{sim.ProbabilitySpace}
\alias{sim.ProbabilitySpace}
\title{Simulate n draws from probability space.}
\usage{
\method{sim}{ProbabilitySpace}(self, n)
}
\arguments{
\item{n}{(int): How many draws to make.}
}
\value{
A vector containing the Simulation results.
}
\description{
Simulate n draws from probability space.
}
