% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variables.R
\name{Sim.RV}
\alias{Sim.RV}
\title{Simulate n draws from probability space described by the random
  variable.}
\usage{
\method{Sim}{RV}(self, n)
}
\arguments{
\item{n}{(int): How many draws to make.}
}
\value{
Results: A vector or matrix containing the simulation results.
}
\description{
Simulate n draws from probability space described by the random
  variable.
}
