% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{tabulate.Results}
\alias{tabulate.Results}
\title{Counts up how much of each outcome there were.}
\usage{
\method{tabulate}{Results}(self, normalize = FALSE)
}
\arguments{
\item{normalize}{(bool): If True, return the relative
frequency. Otherwise, return the counts.
Defaults to False.}
}
\value{
Data frame: A data frame with each of the observed
outcomes and their frequencies.
}
\description{
Counts up how much of each outcome there were.
}
